requirements.txt
aiogram>=3.17.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.30.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.32.0
alembic>=1.14.0
psycopg[binary]>=3.2.0
